Fix EmsRefresh for SwiftManager Caused by Duplicate Method Names
Two issues with EmsRefresh for CloudObjectStores:

1) The code previously living in save_inventory_cloud.rb for CloudObjectStores
was moved to save_inventory_swift.rb.  It should have been *copied* not moved.
A copy was placed back in the original file.

2) The EmsRefresh module is extended by including each of the save_inventory_****.rb files.
Any duplicate method names in the separate files result in an unintended conflict.
When the code above (issue 1) is copied, a conflict in the methods used to save cloud
object store inventory results.  We will modify save_inventory_swift.rb methods in question.